+++ examples/core/thread_cancel.rb
@@ -7,7 +7,7 @@ require 'polyphony/auto_run'
 
 def lengthy_op
   100.times do
-    sleep 0.01
+    orig_sleep 0.01
     @op_count += 1
   end
   @op_count
@@ -15,11 +15,12 @@ end
 
 spin do
   cancel_after(0.1) do
-    data = Polyphony::Thread.spin { lengthy_op }.await
+    data = Polyphony::Thread.process { lengthy_op }
     puts "slept #{data} times"
   end
 rescue Exception => e
   puts "error: #{e}"
 ensure
+  sleep 0.1
   puts "slept #{@op_count} times"
 end