+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Patch Fiber#inspect to show correct block location
+* Add Gyro.run
+* Move away from callback-based timer API (from now on only Timer#await)
+
 0.22 2020-01-02
 ---------------