+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Rewrite `thread`, `thread_pool` modules
 * Add `Kernel#orig_sleep` alias to sync `#sleep` method
 * Add optional resume value to `Gyro::Async#signal!`
 * Patch Fiber#inspect to show correct block location