+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Add support for cancelling multiple coprocesses with a single cancel scope
 * Fix stopping a coprocess before it being scheduled for the first time
 * Rewrite `thread`, `thread_pool` modules
 * Add `Kernel#orig_sleep` alias to sync `#sleep` method